src/music/requirements.txt
numpy
pandas
black
datetime
matplotlib
seaborn
spotipy
python-dotenv
bs4
pyarrow